OMAP4: DSS: Support of user pointer for video playback replication
This patch uses user pointer to do the video playback on two different
display devices simultaneously using two pipelines. In this case, the
tiler allocation happens only once in which the frame data is read by
application. The same buffer is used by the other pipeline by means of
converting the virtual address to physical address of that buffer.

Change-Id: Ica8786d294f098b2550021d70b647ddb5f093e63
Signed-off-by: Sujeet Barnawal <s-baranwal@ti.com>