omap4: usbhs: Enable I/O wakeup during global/suspend
This patch reorganizes the usbhs (EHCI and OHCI) code.
The enableing the functinal clocks and system configuration is
performed irrsepective of functionlaity of the EHCI/OHCI.
This helps the future implemenation of remote wakeup by I/O daisy chain.

The patch enables the I/O daisy when ever the EHCI/OHCI
enter the glaobl suspend. the I/O daisy chain is disabled
once the EHCI/OCHI drivers are resumed.

Change-Id: I794e9a5a88622d0f9c548450c57ec40cff53a3e7
Signed-off-by: Keshava Munegowda <keshava_mgowda@ti.com>