omap4: usbhs: Enable remote wakeup using I/O pads
The remote wakeup implemented using I/O pads of the
USB ports. When there is not device connected to
USB host (EHCI/OHCI) the clocks are disabled in the
bus suspend path. The clocks are enabled when I/O
interrupt triggers. The USBHS clock and port clocks
are enabled using workqueue triggered by PRCM interrupt
handler of I/O.

The common initialization is done as part of USBHS
core driver instead of independent EHCI/OHCI drivers
initialization. The common initialization and exporting
the resouces to EHCI and OHCI reduces the ioremap calls and
simplifies the implementation.

Change-Id: I947488a665ac22927465c4bc0b720162691e6213
Signed-off-by: Keshava Munegowda <keshava_mgowda@ti.com>