syslink: ipu_pm: fix suspend/resume when no IPU image loaded
ipu_pm_drv_suspend and _resume were trying to notify ipu for
suspend/resume without cheking if an image was loaded leading
to an error in the suspend path.

Checking the handle to each proc before sending the notification
is avoiding this.

Suspend/resume is working now with/without an image loaded in IPU.

Change-Id: I55aca1674b1cdffba9fb3247e7da605b4082a521
Signed-off-by: Miguel Vadillo <vadillo@ti.com>