TWL6030-IRQ: wait for irq to be enabled.
When the ac charger is plugged in on suspend,
the irq may status may still be set to disabled
when the threded irq handler is called.
This gives some time for the irq to be enabled by
resume_device_irqs.

this also seems to fix a problem where usb is not
recognized if the cable is attached while on suspend.

[Leed] Fix checkpatch warnings and file change mode

Change-Id: I000cb43a79e0874cd4faa71878930b8dce029d1c
Signed-off-by: Axel Haslam <axelhaslam@ti.com>
Acked-by: Leed Aguilar <leed.aguilar@ti.com>