omap4: pm: Add SAR sar backup support towards device off
The SAR RAM is maintained during Device OFF mode. The register layout
is fixed in SAR ROM. SAR is split into 4 banks with different
privilege accesses based on device type

 ---------------------------------------------------------------
 Access mode		Bank	Address Range
 ---------------------------------------------------------------
 HS/GP : Public		1	0x4A32_6000 - 0x4A32_6FFF (4kB)
 HS/GP : Public		2	0x4A32_7000 - 0x4A32_73FF (1kB)

 HS/EMU : Secured
 GP : Public		3	0x4A32_8000 - 0x4A32_87FF (2kB)

 HS/GP :Secure
 write once.		4	0x4A32_9000 - 0x4A32_93FF (1kB)
 ---------------------------------------------------------------

The save process is done entirely by software and restore is done by
hardware using he auto-restore feature. The restore feature is enabled
by default and cannot be disabled. The software must save the data
to be restored in a dedicated location in SAR RAM.

Change-Id: I8fc8789989d7d31da1d60a68966f3a15988217c4
Signed-off-by: Santosh Shilimkar <santosh.shilimkar@ti.com>