omap4: pm: Add GP device-off support
This patch adds device off support to omap4 GP device type. Because of
ROMCODE BUG, GIC and WAKEUPGEN is not restored. Hence HLOS should
restore these on it's own. Even with this workaround, the device off
mode can't be fully supported on GP devices.

This patch adds the support for completeness point of view and in
case the issue gets fixed in the next silicon version.

By default sw settings, the GP device won't support device off mode

Change-Id: Icb9959fba616d960cb1aa4989f858d4fd970b86e
Signed-off-by: Santosh Shilimkar <santosh.shilimkar@ti.com>