serial: do not throw away MSR deltas when resuming
When resuming, serial_omap_startup is called re-initialise the UART port.
Unfortunately, the MSR values were ignored in this case, and a change e.g. on
CTS was therefore lost (deltas bits of this register are reset on read).

In our case, this causes the UART to not send a message to the modem and the
phone needs a reboot to become operable again.

Change-Id: If85401046f0a29d07202c59fb10f8b2fe910fbb6
Signed-off-by: Herve Fache <h-fache@ti.com>