OMAP2+: DMA: OFF mode support
The DMA registers will lose context in OFF mode which needs to be saved
and restored in suspend/resume path.

All the channel specific registers and global registers which are in use
at the time of suspend will be saved and restored during resume.

Tested on OMAP4430SDP ES2.1 EMU board with McBSP test cases.

Change-Id: I789de2d6e6502bb3dd9d3ab1575312a1fa45d12f
Signed-off-by: G, Manjunath Kondaiah <manjugk@ti.com>
Reviewed-by: Santosh Shilimkar <santosh.shilimkar@ti.com>
Reported-by: Shubhrajyoti Datta <shubhrajyoti@ti.com>
Tested-by: Shubhrajyoti Datta <shubhrajyoti@ti.com>
Acked-by: Leed Aguilar <leed.aguilar@ti.com>