OMAP4: HSI: Fixed scheduling inside spinlocks
The HSI driver was heavily protected with spinlocks, including function calls.
But some of these functions were calling PM runtime code, which in turn calls omap_device code, and there are many semaphore protections in omap_device.
Thus we were facing the BUG : scheduling while atomic, because if the semaphore is taken, the omap_device tries to sleep, which is forbidden inside a spinlock.
So I have remove the spinlock protection around function calls, but kept it around shared data and registers, allowing (I hope) multiple HSI functions to be concurrently called for different channels.

Change-Id: I49d5d13cfbbe672c770c277bbdf6362c7dc3c063
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>