TWL6030 Battery: Update the capacity properly.
For android userspace, we need to update the battery
capacity properly, shoing 100% when battery is full.
Because the "the load" will always consume more current
than the cutoff, BQ will never indicate charge complete,
and battery will always charge. This is a known limitation
of the Circuit configuration.

For this reason, we asume that more than 3.9v is a full battery.

Also while charging The voltage is not battery voltage,
but Voreg, so we force 90% capacity, so that the
android icon will show status as charging, and not
"charged"

Change-Id: I13f4f0d309f098d3a1a19a12eb8c0a9ef8e48450
Signed-off-by: Axel Haslam <axelhaslam@ti.com>
Acked-by: Dan Murphy <dmurphy@ti.com>