usb: musb: Remove idle path phy clock and device enable/disable usage
For core domain transition/wakeup, all the clock and
sysconfig register configurations for musb were done through idle path.
Removed the idle path dependency as we can use the connect/disconnect
events path for doing the same.

Introduced the new bit field SMARTIDLEWKUP for omap4 and using it.
Also as per OMAP4 TRM musb has to be configured to smart idlewkup/standby
mode while it is active and configure it to force idle/satndby when
disconnected. Configuring it smart idle/standby itself will block the
core retention when it is active.

Fixed the issue in phy_clk_set when this api is called back to back
with ON as parameter.

[Moiz] - Fixed the checkpatch warnings

Change-Id: I1686ee21aae775b8fb7f0d80e8970a46fbb3e0c6
Signed-off-by: Hema HK <hemahk@ti.com>