TWL6030: Battery: Wakeup from suspend to refresh watchdog
If ac is charging the battery we need to wake up
at least every 30 seconds to refresh the 32 second
watchdog timer on the BQ chip.

This patch adds the option to prevent suspend
when the battery is charged from ac by enabling
a wakelock:

echo 1 > /sys/class/power_supply/twl6030_battery/device/wakelock_enable

if the wakelock is not enabled, and the battery
is charging through AC, a wakeup is programmed so that
the watchdog can be reset.

Also, this patch shuts down the battery every 10 minutes
so that the actual voltage level of the battery can be
determined and reported to the user while charging.

Change-Id:	I6e81c40f1360b10cf47b35c9ce2915074112edd5
Signed-off-by: Axel Haslam <axelhaslam@ti.com>