omap iommu: Insert a gap page between IOVMAs against override
Inserting a gap page between IOVMAs could detect an override on
other IOVMA with iommu fault. This was originally suggested
by Sakari Ailus and based on the work and comment by David Cohen.

If there's any mmap in the list and the flags has IOVMF_DA_ANON,
start and prev_end can have the same value at the end of the loop.
Without this patchset, we can get "no space to fit.." error.

Change-Id: Ib47456d1668576c3ff6585c6d45d1e39712e4d86
Signed-off-by: Hiroshi DOYU <Hiroshi.DOYU@nokia.com>
Cc: David Cohen <david.cohen@nokia.com>
Cc: Sakari Ailus <Sakari.Ailus@nokia.com>
Signed-off-by: keun-o.park <sahara@ti.com>