OMAP4: Improve consistency of HDMI Attach / Dettach Interrupt
Continous & random attach dettach interrupts were generated
depending on the TV. Debounce register was used to determine
the actual event based inside the workqueue to avoid wrong
notification of event to the user side. This helped to remove
the previous workaround of enabling some internal core interrupts
and the delays added.

Signed-off-by: Mythri PK <mythripk@ti.com>
Signed-off-by: Dandawate Saket <dsaket@ti.com>

Change-Id: Ia91f082512abf6162f3d79779c911247084c8940