ASoC: ABE DSP: Avoid suspend when ABE is active
Suspend is requested by User space to have core domain
and (at least) DSS domain to off.
But it doesn't mean that ABE needs to go to OFF also.
During voice call, CPU cores are not used,
User can request a suspend even ABE is working for modem audio.
Therefore the rule is simple if ABE has an activity,
it doesn't go to sleep even the rest of the system goes.

Change-Id: I2b846dd8c02fde466cd2d5d75b153d2740374a3f
Signed-off-by: Francois Mazard <f-mazard@ti.com>