ASoC: SDP4430: add no_suspend on BE and legacy DAIs using TWL6040
soc_suspend doesn't check if DAIs are shared between non-suspendable
DAI link and suspendable DAI link.
In the voice call usecase, modem DAI link needs to continue to run and
its DAIs shared with other DAI links need also to stay on.
The solution is to put as non-suspendable the other DAI links using the same
DAIs.
In Android, the other non-suspendable DAI links are not on
if a suspend is requested.
This patch is a workaround waiting for a ASoC multicomponent rework ongoing.

Change-Id: Ie74d23d36d28979efd366d4c3df208c282a9e878
Signed-off-by: Francois Mazard <f-mazard@ti.com>