OMAP4: HSI: Workaround for CAWAKE quick transitions
There is an issue with CAWAKE interrupt status bit.
The problem is that 2 different events (CAWAKE rising edge and falling edge) generate the same interrupt.
It is not an issue when 2 different CAWAKE transitions occur with sufficient time for SW to process the interrupt. In this case, the SW can read the CAWAKE line status when the interrupt occurs.
But in the case we have a CAWAKE LOW->HIGH->LOW or HIGH->LOW->HIGH transition occurring very fast, the SW will read the CAWAKE line status when first interrupt occurs, and it will read a wrong value, because CAWAKE line status will have changed already.

This SW workaround adds a SW check of what the CAWAKE line should be at the time of CAWAKE transition, and detects if the CAWAKE value is what is expected (single transition occured) or if we miss one transition (double transition back to back).

Change-Id: I931d76f95dca095f093f3dbbccea601a2c2aec19
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>