OMAP4: HSI: Fix for potential race condition when accessing DMA channel info
There is a potetial race condition occuring when processing DMA interrupts.
As the DMA processing function is not secured by spinlock from the top, there is a chance that another thread (on another CPU) tries to access or free the DMA channel while accessed by the bottom half.
This could cause the "Unable to match DMA channel issue"

Change-Id: I6b42087303e784fc730674a5c8c86014d4b93198
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>