OMAP4: DSS: Fix GFX fps performance with sched_update
This patch fixes the GFX app fps mesasurements from 60 to 240.
when the update was happening the subsequent calls were getting
blocked as it was unable to acquire taal panel device lock.
This was causing the 60fps. Unblocking the panel device lock request
fixed the issue.
The sched_update was not specified for the taal2 as this was causing
the update to be used for 2nd display.
240 fps can now be achieved on both the displays.

Change-Id: I67ff5f56ad480c9d47b0c9773ecb0da7a81f4315
Signed-off-by: Lajos Molnar <molnar@ti.com>
Signed-off-by: srinivas pulukuru <srinivas.pulukuru@ti.com>