OMAP3: DSS: Fix OMAP3 panic due to OMAP4 function being called
Patch I1d10b89e158c4b63235fbd7566a6e249f34528b9 introduced a
number of changes for OMAP4/DSS.  With this patch,
dpi_set_dispc_clk() unconditionally calls the function
dss_select_lcd_clk_source() which is for OMAP4 only and leads
to a panic on OMAP3.

This patch makes that call runtime conditional on OMAP4.

The rest of the changes appear to handle OMAP3
properly and are now tested on OMAP3.

Change-Id: I8b1dd41816185e1f45be029c25adfcd4c857046f
Signed-off-by: Steve Schefter <sschefter@ti.com>
Reviewed-by: Praneeth Bajjuri <praneeth@ti.com>