OMAP4: HSI: PM: changed HSI default clock to 96MHz
In order to ensure HSI will function even at OPP50, we set the HSI Functional Clock to 96MHz instead of 192MHz.
The default functional clock is defined in omap_hsi_platform_data, when HSI device is registered.
Then upon HSI driver registration, we use the defined value by calling the omap_device API (omap_device_set_rate()).

Warning : with this commit, the functional clock is changed, so it might be necessary to adapt the HSI TX and RX divisors.

Change-Id: I90742c621a177128d55b2ce720ffd7a4cd83fb41
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>