OMAP4: Make HDMI attach / dettach consistent with Smart idle mode
HDMI attach detach needed modification for the core to be in Smart idle
wakeup capable and the interrupts need to be set to wakeup capable.
Also delays which were not required needs to be removed. Debounce delay
had to be optimized. Handled the OCP timeout scenario caused due to
L3 idling delay. Increased SCP clock frequency to get better response
from Core.

Patch fix for mixed declaration and code

Moved the struct definition to variable declaration portion
of the code

Change-Id: I7ea76178b4a038cf4305601aa35bc1dffd802cee
Signed-off-by: Dandawate Saket <dsaket@ti.com>
Signed-off-by: Dan Murphy <dmurphy@ti.com>