OMAP4:Fix the glitch on GPIO54/55.
The OMAP4 output gpios seem to have a glitch when OMAP4 comes out
of OFF mode.Workaround to avoid this for GPIO54/55 lines used by
WLAN and BT chips on Blaze.

The errata is not yet confirmed to be applicable to OMAP4.Once
confirmed a generic implementation is required to handle all output
gpios. Revert this patch once a generic patch becomes available.

Change-Id: I530db6d269ad27a282307f91d985b5b4598cac68
Signed-off-by: Madhusudhan Chikkature <madhu.cr@ti.com>