OMAP4: HSI: changed HSI default FIFO mapping
The default HSI FIFO mapping was set to 8 FIFOs per port, which is SSI compatible mode.
This caused the HSI channels [8..15] to not be associated to a valid FIFO.

With the change, the FIFO mapping policy is to map all FIFOs to HSI port 0, so every channel has
its own FIFO.

Note : This fix has been removed from Gerrit by mistake, and needs to be re-introduced.

Change-Id: I6f6ee0b06d0522841a274080661ee2cce49feecf
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>