OMAP4:HSMMC5 context save/restore in IDLE
The HSMMC5 looses context in IDLE when OSWR is hit.Populate an
API to get the status to restore the context.

Remove the L4 PER constraint from the hsmmc driver which prevented
PER domain OSWR.

The HSMMC5 is not connected through regulator. So power_saving flag
makes no sense for HSMMC5. Unset this flag.

Change-Id: Ibfd5b77ce5d24683592dfcb06fb8600fb960cb6f
Signed-off-by: Madhusudhan Chikkature <madhu.cr@ti.com>