OMAP: Serial: Define UART idle timeout in milliseconds
The OMAP-Serial driver has a variable called "idle_timeout" that
defines how long the UART must be inactive before disabling clocks.
This timeout is currently in seconds and hence the minimum amount
of time before the clocks can be disabled is 1 second. Modify the
driver to define the idle timeout in milliseconds to allow finer
tuning on when the clocks can be turned off.

Change-Id: I5abef270d08e4a239b89c1446d573ddaaab4b4e2
Signed-off-by: Jon Hunter <jon-hunter@ti.com>