OMAP: Serial: Rename rx_timeout to rx_poll_rate
When using DMA there are two timeouts defined. The first timeout,
rx_timeout, is really a polling rate in which software polls the
DMA status to see if the DMA has finished. This is necessary for
the RX side because we do not know how much data we will receive.
The secound timeout, RX_TIMEOUT, is a timeout after which the
DMA will be stopped if no more data is received. To make this
clearer, rename rx_timeout as rx_poll_rate and rename the
function serial_omap_rx_timeout() to serial_omap_rxdma_poll().

Change-Id: Ieffcb966375c7e8c026ea151ba9532949811910b
Signed-off-by: Jon Hunter <jon-hunter@ti.com>