OMAP: Serial: Make the RX_TIMEOUT for DMA configurable for each UART
The OMAP-Serial driver defines an RX_TIMEOUT of 3 seconds that is
used to indicate when the DMA for UART can be stopped if no more
data is received. The value is a global definition that is applied
to all instances of the UART.

Each UART may be used for a different purpose and so the timeout
required may differ. Make this value configurable for each UART so
that this value can be optimised for power savings.

Change-Id: Ib6422821d85c83fa6b7d9989d368d92af2016355
Signed-off-by: Jon Hunter <jon-hunter@ti.com>