OMAP: Serial: Make the UART idle timeout configurable
The OMAP-Serial driver hard codes a time of 5 seconds that is used
to determine if a UART is active. In other words, if a UART is
inactive for more than 5 seconds then the clocks for the UART can
be disabled.

Each UART may be used for a different purpose and so the idle time
that determines if a UART is inactive may differ. Make this value
configurable for each UART so that this value can be optimised for
power savings.

Change-Id: I15f7278bbe9d81d39d1cd6c48a77be2cf9a288db
Signed-off-by: Jon Hunter <jon-hunter@ti.com>