TWL6030: BATTERY: Fix charging code in case device attached
With this patch, we check if the MUSB OTG module is switched
to HOST mode (ID GND) when a USB device is connected to the
OTG port and do not enable USB charging.

Also it was observed that removing the USB device was causing
a glitch in the VBUS and hence was generating VBUS_OVP fault,
which is not correct as its in HOST mode and is actually not
getting charged by an external USB charger. This issue is
fixed by Masking the MCHARGERUSB_FAULT interrupt when a device
is connected and Unmasking the interrupt when the device is
detached.

Change-Id: I2e199382b356e3d46875acef82dd82c1b0460237
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>