OMAP4: Avoid calling clk_get() from within local_timer_setup()
The function local_timer_setup() is called from within the context
of a spin-lock. Currently the local_timer_setup() function is
calling the clk_get() which in turn uses a mutex which may sleep.
Functions that may sleep should not be called from within the
context of a spin-lock and so modify the local_timer_setup()
to avoid this call to clk_get().

Change-Id: Ib2bee47eb69aebf652123ce42bb6f3c182c5f800
Signed-off-by: Wenbiao Wang <wwang@ti.com>
Signed-off-by: Jon Hunter <jon-hunter@ti.com>
Acked-by: Santosh Shilimkar <santosh.shilimkar@ti.com>