OMAP: UART:Allow UART ports to declare a wakelock or not.
Devices may have a need to hold a UART wakelock for
individual ports.  Thereform moving the wakelock code
to the board file so that the wakelock can be configured
based on the UART number and not hard coded.

If a wakelock needs to be held for a specific UART then
the port should set the function pointer.  If it is not
needed then the function pointer should be NULL.

Change-Id: Ifc4e598e5bfcca0e26a15b8c12d22d02debf8a5c
Signed-off-by: Axel Haslam <axelhaslam@ti.com>