OMAP4: HDMI: Workaround for ACR packets - part 1
This workaround is used to send ACR packtes to HDMI at the rate
specified by the test 7-29 of the HDMI Compliance Test
Specification. It uses the Notify module of Syslink to send an
event to SYSM3 indicating the window size during which a single
ACR packet should be sent. This time is specific of the video
resolution that is selected.

Change-Id: Ic0e0bbdbb2102a5e772b9d86b862ef8fc5054700
Signed-off-by: Ricardo Neri <ricardo.neri@ti.com>