V4L2 changes for controlling z-order
Description# v4l2 window structure is extended to take z-order for the
window.
The default assignment for the video pipes is modified to assign the
highest value (zorder 3) for graphics plane and lowest for video pipes.
Overlay format set and get are modified to configure the z-order.

Validation# AV playback on overlay with UI Controls on graphics plane

Change-Id: I097a45f6b398e33d546ff5baba05ee9eb32ed3fc
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>