Nohz/sched: don't go tickless when CPU is loaded
One criterion of C-states selection is based on the load factor.
High load prevents deep C-states.The load is evaluated and
updated at each scheduler tick.

During the time when NOHZ mode is active the evaluation of the
load will not be accurate. So when a high load is evaluated on
a tick happening on a burst of activity, this load value is kept
until the next tick, which can take few ms to happen in NOHZ mode.
Therefore, this patch ensures that the load is taken into
consideration before going tickless, so that the load-based
calculations done in CPUidle are done using an accurate value.

Delaying NOHZ decisions until the load is zero, improved the load
estimation on our ARM/OMAP4 platform where HZ =128 and increased
the time spent in deep C-states (~50% of idle time in C-states
deeper than C1).A power saving of ~20mA at battery level is
observed during MP3 playback on OMAP4/Blaze board.

Change-Id: Ib94e2e94ade843fd0500f8484aac2f88722b6e8d
Signed-off-by: Nicole Chalhoub <n-chalhoub@ti.com>
Signed-off-by: Vincent Bour <v-bour@ti.com>
Acked-by: Akash Choudhari <akashc@ti.com>