OMAP4: HSI: PM: added context SAR mechanism
The runtime PM .runtime_suspend() and .runtime_resume() callbacks now call the HSI save and restore context functions.
Thus the HSI context is saved and restored not only when entering OFF mode, but also when entering IDLE and RET modes.
This is sub-optimal, but aligned with OMAP PM chosen policy.

Also at boot time, the CAWAKE line is configured to be wake-up capable from OFF mode.

Change-Id: I8e3687efdebb571e2dd8989a78bd34df97d27420
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>