OMAP3: dss: suspend/resume fixes
OMAP3 call to disable the clocks in dss_mainclk_disable
meant that the DSS clocks were not turned off and the
DSS power domain never went idle.

The panel uses the DPI interface which has a disable
but not a suspend option.  As a result, the display is
placed into DISABLED, not SUSPENDED.

So we need to allow for that state when resuming.

Change-Id: I53f37f770cd79d41b544cfadd7aee142573594c6
Signed-off-by: Steve Schefter <sschefter@ti.com>