OMAP3: DSS: resolve earlier work-arounds
The initial OMAP 3 patch included a couple of FixMes. This checkin fixes
those. The structure type of the data poited to by platform_data is not
consistent for all devices. The correct one is used and later use of pdata
comes from core.pdata rather than the varying type pdev->dev.platform_data.
Also, during init, core.pdata will not yet be initialized.  It is therefore
necesssary to check that before deciding to save the context when enabling
or disabling clocks.

This change is similar to Ie17c30959782f584ec697fd9876ef6bbbb44d0f7 which was
later reverted due to a panic in an enviroment with PM turned on.  That is
now resolved by checking core.pdata rather than changing the order of the
platform_driver_register() calls (an therefore the order that the probe
routines run).

Change-Id: I4996f5d8305e30642bf95d5dbb9ddac7c2446519
Signed-off-by: Steve Schefter <sschefter@ti.com>