syslink: ipu_pm: fix disable/enable in runtime ipu hib
Adapt IPU Hibernation override control flag to work
with the new watch-dog timer design.

In order to enable/disable the flag hibernateAllowed should
be set to 0 and hibernation and wdt will be disabled.
Set it the flag back to 1 will allowed again both features.

This flag is located in the member pm_flags of the sms struct
and it is the last bit of it.

Very useful when debugging ipu and avoid the rebuilding
of the kernel after disabling the features in menuconfig.

Change-Id: I4b58bb62912c4481d068b347d2feadc9fe95f3fc
Signed-off-by: Miguel Vadillo <vadillo@ti.com>
Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>