OMAP: PM: runtime_idle not called for platform drivers
The platfom bus implementation of PM_Runtime (arch/arm/mach-omap2/pm_bus.c) doesn't call
the runtime_idle() method of the platform driver before queuing a suspend request for the device.
This causes the device to be suspended unconditionnaly, without giving the driver a chance to check
if the device can be suspended.

This contradicts the pm_runtime principles, where the bus runtime_idle() method should :
"check if the device can be suspended (i.e. if all of the conditions necessary for
suspending the device are satisfied) and to queue up a suspend request for the
device in that case."
Cf kernel/Documentation/runtime_pm.txt

Change-Id: Id13d1654d247cadb4c824479ed5f0f1c4ca2f0ea
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>