OMAP4: HSI: Replace pm_runtime_xxx usage by omap_device_xxx usage
The PM_Runtime APIS are not callable from an atomic context, because they can possibly call
the schedule() function, which is not allowed inside an atomic context.
But the HSI driver must enable the clocks in the HSI bottom half handler (ie. a tasklet) when an interrupt is received.
Consequently, we cannot enable the clocks inside the tasklet, and would be forced to use a workqueue, which a much higher latency.
Additionnaly, when wakeup from OFF mode occurs due to an external HSI event, we must enable the HSI clocks from
PRCM SW code, and if we call pm_runtime API from here, we might schedule() while almost all interrupts are disabled, which
is probably not a good thing.

This is why I have no simple choice but bypass the pm_runtime API, until this API is made "Atomic context" friendly.

Change-Id: Iaba89f2c38b744868598541843f52f9e8976c89b
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>