arm: omap4: board-4430sdp update for BTFM
TI ST driver for shared transport on UART and BTWILINK
for the Bluetooth core on the WiLink chip are platform
drivers.
Generally UART devices have no way to be probed and don't tend
to have platform data unless the BSP/bootloader send them via
SFI/ATAGS.
On other platforms this was the requirement, and hence ST driver
has the platform data which populates UART details.

Change-Id: I9db10829caaa602dffd8223005e67b74d433ea2f
Signed-off-by: Pavan Savoy <pavan_savoy@ti.com>