TWL6030: USB: Allign twl-irq code with open-source
This patch alligns the twl-usb irq handler code with the open
source code.

This new code fixes a BUG wherein in HOST-only mode, in cold
boot case, the device is not enumerated as the ID_GND bit
of USB_ID_INT_SRC register is set after a delay of approx
7 msec from ldo_init() where ID comparator is enabled in TWL
init path. Hence instead we use STS_HW_CONDITIONS register
to detect the presence of a device during cold-boot.

Change-Id: Iec8182847b1eca4dcd372e127e53ea40195344d7
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>