USB: MUSB: Move wakelock/L3 constraints to musb irq handler
This patch moves the wake_lock/unlock calls and hold/release
L3 constarint calls on attach/detach of USB cables from
musb_notifier calls in TWL irq path to musb_irq handlers
in the CONNECT(HOST)/RESET(GADGET) and DISCONNECT interrupts
in musb stag0 irq handler.

This patch is strongly dependent on the commit:
1d4ed7b251a13f749fca0639d5cad575cafd981f

With this above patch the cold-boot in HOST only mode is
fixed but when it tries to hold and L3 constarint HWMOD
is not yet initialised and hence falls into a timing issue.
Moving the L3 constarint calls to musb irq handler solves
this problem.

Keeping the c state constarint calls in the musb notifier
as cannot move this to irq_handler. Adding the constraints
data to musb platform data. Also, adding the missing
release of c state constraint on a detach.

Change-Id: If3c7bb83226293d7216a97856c734778c1457366
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>