OMAP4: HDMI: PM Cut 48 MHZ option call when HDMI is not connected
In current solution HDMI optional clock is couple with DSS clocks.
All the clocks are turned on and off in main clock function.
So if DSS is active and no HDMI connected still the optional clock
is kept active. This is not allowing the DSS domain to hit retention.
Separated out the optional clocks from the dss clocks. This clock is
still required to make sure the HDMI PLL is locked and when the HDMI
PHY is full power mode i.e the HDMI active during use case.

Change-Id: Iad41a40565f970d9cb12111f00fe51e3d472442e
Signed-off-by: Dandawate Saket <dsaket@ti.com>