OMAP4: DSS: Fix for interlace video rotation
In the display controller plane setup, the height and width for the
interlaced videos is set without considering the rotation values.
Because of this, 1920*1080i videos were not playing properly when the
rotation is set to 90 or 270 degrees.

This patch sets the height and width values based on rotation.
Tested on Blaze ES2.1 with different resolution videos and different
values of rotation.

Change-Id:I7d28e5c0e63fd26b274139481e5b7ef40baf8b04
Signed-off-by: Parasuraman Ramalingam <x0151836@ti.com>
Signed-off-by: Srinivas Pulukuru <srinivas.pulukuru@ti.com>