OMAP4: HSI: Avoid spurious HSI Fclk transition fo 192MHz
The HSI FClk can be 96MHz or 192MHz, and defines the RX/TX link speed between OMAP and Modem.
Thus once set by upper layers (through omap_device_set_rate() API), HSI Fclk shall not change when Core OPP changes.
Furthermore, 192MHz TX/RX rate can only be achieved at OPP100, and shall force the core voltage domain to 1.1V and prevent any transition to OPP50.

HSI layer calls omap_device_set_rate(96Mhz) to force the HSI clock to 96MHz.
omap_device framework checks what is the voltage associated with 96MHz, and founds it is 0.93V.
But if another device of the same voltage domain has requested a 1.1V because it has to run at OPP100, then omap_device sets all the devices of the voltage domain to 1.1V, and to the frequency associated to 1.1V, which is 192MHz for HSI.

So to summarize, we have requested HSI to run @96MHz, and because another device needs to run at OPP100, then we force the HSI to be also @OPP100 andthe HSI Fclk to be @192MHz.

A temporary workaround is to define all HSI OPP to be @96MHz, whatever the voltage.

Change-Id: I136c6dd946cdbc505ca63475a16a61129ca2224a
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>